fix(web): address tablet preview review feedback
