+++ apps/web/lib/utils/file-types.test.ts
@@ -1,0 +1,13 @@
+import { describe, expect, it } from "vitest";
+import { isMarkdownFile } from "./file-types";
+
+describe("isMarkdownFile", () => {
+  it("recognizes Markdown extensions case-insensitively", () => {
+    expect(isMarkdownFile("docs/README.md")).toBe(true);
+    expect(isMarkdownFile("docs/README.MDX")).toBe(true);
+  });
+
+  it("rejects non-Markdown files", () => {
+    expect(isMarkdownFile("src/index.ts")).toBe(false);
+  });
+});
